Issue #12101: move global locale to localized message
