+++ src/main/java/com/puppycrawl/tools/checkstyle/DefaultLogger.java
@@ -24,7 +24,6 @@ import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.io.Writer;
 import java.nio.charset.StandardCharsets;
-import java.util.Locale;
 
 import com.puppycrawl.tools.checkstyle.api.AuditEvent;
 import com.puppycrawl.tools.checkstyle.api.AuditListener;
@@ -161,7 +160,7 @@ public class DefaultLogger extends AutomaticBean implements AuditListener {
     public void addException(AuditEvent event, Throwable throwable) {
         synchronized (errorWriter) {
             final LocalizedMessage exceptionMessage = new LocalizedMessage(
-                    Definitions.CHECKSTYLE_BUNDLE, Locale.getDefault(), DefaultLogger.class,
+                    Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.class,
                     ADD_EXCEPTION_MESSAGE, event.getFileName());
             errorWriter.println(exceptionMessage.getMessage());
             throwable.printStackTrace(errorWriter);
@@ -171,7 +170,7 @@ public class DefaultLogger extends AutomaticBean implements AuditListener {
     @Override
     public void auditStarted(AuditEvent event) {
         final LocalizedMessage auditStartMessage = new LocalizedMessage(
-                Definitions.CHECKSTYLE_BUNDLE, Locale.getDefault(), DefaultLogger.class,
+                Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.class,
                 AUDIT_STARTED_MESSAGE);
         infoWriter.println(auditStartMessage.getMessage());
         infoWriter.flush();
@@ -180,7 +179,7 @@ public class DefaultLogger extends AutomaticBean implements AuditListener {
     @Override
     public void auditFinished(AuditEvent event) {
         final LocalizedMessage auditFinishMessage = new LocalizedMessage(
-                Definitions.CHECKSTYLE_BUNDLE, Locale.getDefault(), DefaultLogger.class,
+                Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.class,
                 AUDIT_FINISHED_MESSAGE);
         infoWriter.println(auditFinishMessage.getMessage());
         closeStreams();
